Match iTunes previous() method with expected behaviour.
This change makes the previous() method match the behavior of the
spotify extension. Using 'back track' repositions the player to the
beginning of the current track or changes to the previous track if
already at the start of the current track. 'previous track' repositions
the player to the previous track on the playlist regardless of the
position in the current track.