window_filter: track applications by PID
The `apps` and `pendingApps` lists in `window_filter.lua` are indexed by
app name. The window filter will reject a new app if it has the same
name as an existing app. This can be an issue because there's nothing in
MacOS that prevents two apps from having the same name. For example, in
Firefox, if you select "Launch profile in new browser", Firefox will
open a new Firefox application with the same name, bundleID, and path.
Two separate Firefox icons in the Dock will confirm that these are
unique applications.

Switch the window filter to track applications by PID instead of name.
Fetch the application PID in `startAppWatcher()` and stash it in the
`App` object to reference later. Change the function signature of
`windowEvent()` and `appWindowEvent()` to receive a PID instead of app
name and lookup the `App` object by PID.

Both Firefox applications are recognized and tracked separately now. No
regressions in window filter event processing have been observed.